Generate appropriately-sized photos in the output.
These are sized to the exact sizes needed to render these images with
the current stylesheet on a desktop browser, but later it might be best
to scale these up a bit so they'll look nice on a high-resolution display.
The CSS takes care of making it fit even if the bitmap is bigger than it
needs to be.